- fix type;

+++ Market-Analysis/include/imt4reader.h
@@ -52,16 +52,16 @@ public:
     ~IMt4Reader();
 
 protected:
-    QString         fileName;
-    bool            fileExists;
-    FileType        fileType;
-    Header          *header = 0;
+    QString                     fileName;
+    bool                        fileExists;
+    FileType                    fileType;
+    Header                      *header = 0;
     QList<std::vector<double> > *history = 0;
 
 public:
     void setFileName(const QString fName);
     QString getFileName() const;
-    size_t getHistorySize() const;
+    qint32 getHistorySize() const;
     qint32 getHistoryVersion() const;
 
     virtual bool readFile() = 0;
